Add ability to change waiting times via a dialog.
This change adds another option that appears when auto play is enabled.
 The dialog allows user to change all the various wait times.

A dialog class and widget class was added to achieve this.

All the times are now stored in a struct, this struct is then passed
onto the dialog which will modify it, and will be saved if the user
presses ok.

This struct lives in this new dialog class, preferably this would be in
a shared file amongst the widgets.

There was an issue with a connect line in the new widget, currently a
workaround has been in place.