Fix bug - timing issue.
Always stop the autoplay timer when handling a board click.  This will
also prevent the scenario when the auto play animation has arrived at
the next clue but starts a timer to show that clue.  If user clicks on
another clue while that timer is running, the game gets into a bad
state.