Adds a deploy script and all needed resources
Creates a deploy script that will take the app bundle Qt Creator
produces and adds all the needed resources and adds the Qt libraries to
app.  The script will output a ready-for-user app.

This also adds a setup.sh script to run when first setting up the
project so the Qt created app bundles will have the proper resources.

However the user needs to run chmod u+x setup.sh /deploy.sh to be able
to run the scripts.  It maybe useful to write these scripts in python
to avoid having to change permissions.