Introducing StateHandlers
The GamePaneWidget no longer holds an instance to a JeopardyGame
instead it has been abstracted away via the use of a IStateHandler
interface.  This interface takes in the current state and some data (
the clicked table view model index ) and sends out the corresponding
signal containing the next state.

The UI is calls OnStateChanged whenever user makes a change that will
require to update the state.  And UI simply listens to a state changes
and then updates itself to draw that state.

This is the big first refactor step so that the UI can be used for both
offline and online modes.  Which will abstracted by the IStateHandler.

There will be several more upcoming refactoring changes.

This also renames utility to qtutility.