Push 'State Machine' logic down to backend
The UI sends down a StateAction struct and the backend responds with a
StateResponds. So almost all game logic is now done in the backend.

This should now lead to a much smoother transition to the online server
play.  The state handler and server will just need to add the ability
to send those structs across the wire as byte arrays.