Server/Client can recognize when both players are connected.
The online game can now reach the point when both players see the start
game button ( meaning server realized two players had connected).  Then
both players can click the start button and see the next screen.

The server and client now communicate in proper string-yfied
StateActions and StateResponses.

This fixes a bug in toString method.  For some states the row and
column will both be -1.  This is a legal, so add a condition to test
this case.  This change comes with a few modified unit tests.

There is a known bug where server will crash or no longer work when it
closes after establishing a connection, restarting, and having a client
attempt to connect.