Server now handles picking a game.
OnlineStateHandler is refactored a bit more so that the proper classes
handle server responses.

Server-side some logic is added (an extra mode/action) so that the
client identifies which game it wants to play.

Jeopardy online works as before.

There will need to be future refactoring to properly handle launching
each game.